fix(mem_cache+model_runner): address review batch (#25/#29/#31/#33/#34)
Three review-driven fixes folded into one commit:

- (#25) ReplaceTokenPool budget calc: replace _per_token_mha_kv_bytes with
  self._compute_cell_size() — the old helper used TP-total head count +
  raw num_hidden_layers, overestimating available_bytes by tp_size× and
  silently inflating state_max_reqs. _compute_cell_size handles per-device
  heads, align128, adjust_layer_num, and MLA latent fast path; it is the
  project's single source of truth for per-token KV bytes.

- (#29 / #33) ReqToTokenPool.free now clears req.req_pool_idx = None to
  prevent a stale slot index from leaking to later callers; restores
  symmetry with free_chunked / free_recurrent_cache which already cleared
  their respective idx. free_chunked simplified to a self.free(chunked_req)
  delegate; HybridReqToTokenPool.free_chunked override (no-op for
  cross-chunk slot reuse) unchanged.

- (#31) Trim HybridReqToTokenPool docstring (22 lines -> 9 lines).

- (#34) Extract 8 module-level hybrid recurrent helpers (build_hybrid_pools,
  compute_per_req_bytes, split_state_kv_budget, etc.) from model_runner.py
  to a dedicated python/sgl_jax/srt/model_executor/hybrid_recurrent_utils.py.
  Pure relocation: no logic / signature changes. model_runner.py now stays
  focused on the core forward / profile / init flow.

Test imports updated; 14-suite pytest passes (203 cases).
